% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conformal.R
\name{genData.clustered.conformal}
\alias{genData.clustered.conformal}
\title{Generate data for ReadMe categorical example with protected groups}
\usage{
genData.clustered.conformal(
  seed,
  N,
  numgroup = 10,
  numcluster = 5,
  outcome.type = c("Continuous", "Categorical"),
  beta.S = 10,
  alpha.r = 9/10
)
}
\description{
Generate data for ReadMe categorical example with protected groups
}
