% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conformal.R
\name{genData.conformal}
\alias{genData.conformal}
\title{Generate data for ReadMe continuous example}
\usage{
genData.conformal(
  seed,
  N,
  outcome.type = c("Continuous", "Categorical"),
  beta.S = 1,
  alpha.r = 3/4
)
}
\description{
Generate data for ReadMe continuous example
}
