% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conformal.R
\name{get.metric}
\alias{get.metric}
\title{Compute the metrics to evaluation the efficiency and fairness}
\usage{
get.metric(rst)
}
\description{
Compute the metrics to evaluation the efficiency and fairness
}
