% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSurrConformalDR.R
\name{NoSurrConformalDR}
\alias{NoSurrConformalDR}
\title{No Surrogate-assisted Conformal Inference for Efficient Individual Causal Effect
Estimation}
\usage{
NoSurrConformalDR(
  df,
  train.idx,
  eval.idx,
  outcome.type = c("Continuous", "Categorical"),
  SL.library = c("SL.glm"),
  alphaCI = 0.05,
  nested = TRUE
)
}
\description{
See details in \code{\link[=SurrConformalDR]{SurrConformalDR()}}.
}
