% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrClusterConformalDR.R
\name{SurrClusterConformalDR}
\alias{SurrClusterConformalDR}
\title{Surrogate-assisted Clustered Conformal Inference for Efficient Individual Causal Effect
Estimation}
\usage{
SurrClusterConformalDR(
  df,
  train.idx,
  eval.idx,
  numcluster = "auto",
  minsize = 150,
  qvec = c(0.5, 0.6, 0.7, 0.8, 0.9),
  outcome.type = c("Continuous", "Categorical"),
  SL.library = c("SL.glm"),
  alphaCI = 0.05,
  nested = TRUE
)
}
\arguments{
\item{numcluster}{The number of cluster for the non-conformity scores. The default is 'auto'.}

\item{minsize}{The minimum size of each cluster if numcluster = 'auto'.}

\item{qvec}{The selected quantiles for the non-conformity score embeddings.

See details of other parameters in \code{\link[=SurrConformalDR]{SurrConformalDR()}}.}
}
\value{
when \code{outcome.type = "Continuous"}:
\itemize{
\item lower.Y, upper Y: prediction regions for the observed outcomes
with index \code{eval.idx}
\item lower.tau, upper.tau: prediction regions for the individualized treatment effects
with index \code{eval.idx}
}

when \code{outcome.type = "Categorical"}:
\itemize{
\item sets_observed: prediction sets for the observed outcomes
with index \code{eval.idx}
}
}
\description{
Surrogate-assisted Clustered Conformal Inference for Efficiency and Fairness
}
