% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrConformalDR.R
\name{SurrConformalDR}
\alias{SurrConformalDR}
\title{Surrogate-assisted Conformal Inference for Efficient Individual Causal Effect
Estimation}
\usage{
SurrConformalDR(
  df,
  train.idx,
  eval.idx,
  outcome.type = c("Continuous", "Categorical"),
  SL.library = c("SL.glm"),
  alphaCI = 0.05,
  nested = TRUE
)
}
\arguments{
\item{df}{data frame contains
\itemize{
\item X.1, X.2, ...: covariates
\item D: data origin indicator, a binary vector
with 1 being the source data and 0 being the target data
\item A treatment indicator, a binary vector
\item S.1, S.2, ...: surrogates
\item Y: observed primary outcomes
}}

\item{train.idx}{index of \code{df} for model training}

\item{eval.idx}{index of \code{df} for evaluating conformal inference}

\item{outcome.type}{the type of primary outcomes.}

\item{SL.library}{Either a character vector of prediction algorithms
or a list containing character vectors. See details in \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner()}}.
The default is \code{SL.glm}}

\item{alphaCI}{confidence level. The default is 0.05.}

\item{nested}{Logical. Should nested conformal inference be performed
when the primary outcomes are missing. Used only when \code{outcome.type = "Continuous"}.
The default is TRUE.}
}
\value{
when \code{outcome.type = "Continuous"}:
\itemize{
\item lower.Y, upper Y: prediction regions for the observed outcomes
with index \code{eval.idx}
\item lower.tau, upper.tau: prediction regions for the individualized treatment effects
with index \code{eval.idx}
}

when \code{outcome.type = "Categorical"}:
\itemize{
\item sets_observed: prediction sets for the observed outcomes
with index \code{eval.idx}
}
}
\description{
Surrogate-assisted Conformal Inference for Efficient Individual Causal Effect
Estimation
}
